==> Analysis/01.dataset.Rmd <==
---
title: "01_dataset_creation"
author: "Claudia Jiménez"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
 output:
   html_notebook:
    code_folding: hide
    number_sections: yes
    theme: flatly
    toc: yes
    toc_float: yes
---
# Environment

```{r message=FALSE, warning=FALSE}
library(bigrquery)
library(DBI)
library(dplyr)
library(summarytools)
library(tidyverse)
library(forcats)
library(tidyr)
library(DiagrammeR)
```

# BigQuery related functions

This chunks creates the run_query and getSQL function.

```{r}
# Function that takes in a SQL command and runs it on BigQuery. This avoids the connection data in each iteration
run_query<-function(query){
  query_output<-dbGetQuery(con,query)
  return(query_output)
}

# Function for reading SQL files from a folder
getSQL <- function(filepath) {
  con = file(filepath, "r")
  sql.string <- ""
  while (TRUE) {
    line <- readLines(con, n = 1)
    if (length(line) == 0) {
      break
    }
    line <- gsub("\\t", " ", line)
    if (grepl("--", line) == TRUE) {
      line <- paste(sub("--", "/*", line), "*/")
    }
    sql.string <- paste(sql.string, line)
  }
  close(con)
  return(sql.string)
}
```

## Setting up connection

```{r}
bigrquery::bq_auth() # UNCOMMENT WHEN HAVEN TOCKEN ISSUES!!!
# Establecer la conexión a BigQuery
project_id <- "extended-altar-382009"

con <- dbConnect(
  bigrquery::bigquery(),
  project = project_id
)

```

# Data Extraction

Loading queries and extracting the data.
```{r}
apacheadmission <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/apacheadmissiondx.sql"))
patients_emergency <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/patient_emergency.sql"))
patients_admit <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/patients_admit.sql"))
patients_total <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/patient_total.sql"))
apachescore <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/apachescore.sql"))
sensitivitylevel <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/sensitivitylevel.sql"))
charlson_score <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/charlson_score.sql"))
history_dx <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/history_dx.sql"))
pivoted_med <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/pivoted_med.sql"))
shock_vassop <- run_query(getSQL("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Analysis/sql/shock_vassop.sql"))
```

# Joining 1º df

```{r}
#Unite df 
dfs_to_join <- list(patients_total,
                    patients_admit,
                    apacheadmission,
                    patients_emergency,
                    apachescore,
                    sensitivitylevel,
                    charlson_score,
                    shock_vassop) #history_dx

# Perform the left join
final_df <- Reduce(function(x, y) merge(x, y, by = "patientunitstayid", all.x = TRUE), dfs_to_join)

##Exclusion criteria
a <- nrow(final_df)
print('Initial number of patients: ')
print(a)
```
# Exclusion criteria

```{r}
# Age
final_df$age <- as.numeric(gsub("> 89", "90", final_df$age))
final_df <- final_df[final_df$age >= 16,]

##Exclusion criteria
print('Patients >=16 ')
b <- nrow(final_df)
print(b)

# Hospitaladmitoffset = er_los
# Name change, conversion from minutes to hours and absolute values
final_df <- final_df %>%
  mutate(er_los_hrs = abs(hospitaladmitoffset/60)) %>%
  select(-hospitaladmitoffset)

# Remove very high values considered as outcome
final_df <- final_df %>%
  filter(er_los_hrs <= 24)

print('Patients in emergency department during 24 hrs')
c <- nrow(final_df)
print(c)

# Sex
final_df <- final_df %>%
  rename(sex = gender) %>%
  mutate(sex = ifelse(sex != "Male" & sex != "Female", NA, sex)) %>%
  drop_na(sex)

##Exclusion criteria
print('Patients with sex known: ')
d <- nrow(final_df)
print(d)

# Hospital admit source (emergency department)
# DELETE NA
final_df <- final_df %>%
  drop_na(hospitaladmitsource)

## Exclusion criteria
print('patient with emergency department')
e <- nrow(final_df)
print(e)

# Unitstaytype
# DELETE NA
final_df <- final_df %>%
  drop_na(unitstaytype)

## Exclusion criteria
print('patient with unit stay type admit')
f <- nrow(final_df)
print(f)

# Apache admission dx
# Delete rows without values, NA
final_df <- final_df %>% 
  drop_na(apacheadmissiondx)

## Exclusion criteria
print('Patients with apache admission dx known')
g <- nrow(final_df)
print(g)

write.csv(final_df, file = "C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Figures/final_df_sinimputar.csv", row.names = FALSE)

```

# Imputations 

### Sex

```{r}
# Change Male to M and Female to F
final_df <- final_df %>%
  mutate(sex = ifelse(sex == "Male", "M", ifelse(sex == "Female", "F", sex)))

```

### Apache admission dx

```{r}
# Under the same name: pneumonia, bacterial and pulmonary sepsis = sepsis, pneumonia
final_df <- final_df %>%
  mutate(apacheadmissiondx = ifelse(apacheadmissiondx %in% c("Pneumonia, bacterial", "Sepsis, pulmonary"), "sepsis, pneumonia", apacheadmissiondx))

# Apache admission variable, group into the 5 categories with the highest value and the rest grouped into the other variable (this is done due to the small number of many of these diagnostic categories)

# Count the frequency of each category
apache_counts <- final_df %>%
  count(apacheadmissiondx, name = "Frequency")

# Identify the 5 most frequent categories
top_categories <- apache_counts %>%
  top_n(n = 5, wt = Frequency) %>%
  pull(apacheadmissiondx)

# Group less frequent categories under "Other"
final_df <- final_df %>%
 mutate(apacheadmissiondx = fct_lump(apacheadmissiondx, n = 5, other_level = "Other"))

# Count the frequency again after grouping
top_apacheadmissiondx <- final_df %>%
  count(apacheadmissiondx, name = "Frequency")

# Show the result
print(top_apacheadmissiondx)

# List of the 10 categories with the most values
top_categories <- c("sepsis, pneumonia",
  "Sepsis, GI",
  "Sepsis, cutaneous/soft tissue",
  "Sepsis, renal/UTI (including bladder)",
  "Sepsis, unknown",
  "Other")

# Create a new variable "apacheadmissiondx_recoded" that recodes the categories not included in the ranking as "Other"
final_df <- final_df %>%
  mutate(
    apacheadmissiondx = dplyr::recode(
      apacheadmissiondx,
      #"Sepsis, cutaneous/soft tissue" = "Other",
      "Pneumonia, other" = "Other",
      "Pneumonia, aspiration" = "Other",
      "Renal infection/abscess" = "Other",
      "Cellulitis and localized soft tissue infections" = "Other",
      "Sepsis, other" = "Other",
      "Sepsis, gynecologic" = "Other",
      "Infection/abscess, other surgery for" = "Other",
      "Abscess/infection-cranial, surgery for" = "Other",
      "Cellulitis and localized soft tissue infections, surgery for" = "Other",
      "GI perforation/rupture, surgery for" = "Other",
      "Abscess, neurologic" = "Other",
      "GI perforation/rupture" = "Other",
      "Complications of previous open-heart surgery, surgery for (i.e. bleeding, infection, mediastinal rewiring,leaking aortic graft etc.)" = "Other",
      "Complications of previous GI surgery; surgery for (anastomotic leak, bleeding, abscess, infection, dehiscence, etc.)" = "Other",
      "Diverticular disease, surgery for" = "Other",
      "GI obstruction, surgery for (including lysis of adhesions)" = "Other",                                                                       
      "ARDS-adult respiratory distress syndrome, non-cardiogenic pulmonary edema" = "Other",                                                         
      "Thoracotomy for thoracic/respiratory infection" = "Other",                                                                                    
      "Endocarditis"  = "Other",                                                                                                                      
      "Pneumonia, viral" = "Other",
      "Cholecystectomy/cholangitis, surgery for (gallbladder removal)"  = "Other",                                                                    
      "Arthritis, septic"  = "Other",                                                                                                               
      "Meningitis"  = "Other",                                                                                                                     
      "Complications of previous open heart surgery (i.e. bleeding, infection etc.)"  = "Other",                                                     
      "Fistula/abscess, surgery for (not inflammatory bowel disease)" = "Other",                                                                      
      "Sepsis, gynecologic" = "Other",                                                                                                                
      "GI Abscess/cyst-primary, surgery for"  = "Other",                                                                                              
      "Pancreatitis, surgery for" = "Other",
      "GI abscess/cyst"  = "Other",                                                                                                                   
      "Peritonitis" = "Other",                                                                                                                      
      "Peritonitis, surgery for"   = "Other",                                                                                                         
      "Pneumonia, fungal" = "Other",                                                                                                                  
      "Pneumonia, parasitic (i.e., Pneumocystic pneumonia)" = "Other",                                                                                
      "Inflammatory bowel disease, surgery for" = "Other",                                                                                            
      "Myositis, viral"  = "Other" 
    )
  )
```

### Ethnicity

```{r}
# Change empty string = other
final_df <- final_df %>%
  mutate(ethnicity = ifelse(ethnicity %in% c("", "other", "Other/Unknown", "Native American", "Asian"), "other/unknown", ethnicity))
```

### Stratify: Unitdischargestatus (estado del que sale de la unidad)

```{r}
final_df <- final_df %>%
  mutate(
    unitdischargestatus = ifelse(
      unitdischargestatus == 'Expired', 1,
      ifelse(trimws(unitdischargestatus) == '' | unitdischargestatus == 'Alive', 0, NA)
    )
  )

```

### Apachescore,change name and na = median

```{r}
final_df <- final_df %>%
  mutate(apachescoreIV = ifelse(is.na(apachescoreIV) | apachescoreIV == -1, median(apachescoreIV[apachescoreIV != -1], na.rm = TRUE), apachescoreIV))
```

##Treatments

```{r}
final_df <- final_df %>%
  mutate(first_treatmentoffset = replace_na(first_treatmentoffset, 0)) 

final_df <- final_df %>%  
  rename(vasopressors = has_treatment) %>%
  mutate(vasopressors = replace_na(vasopressors,0))
```

# Final df

```{r}
final_df_er_sepsis <- final_df %>%
  select(patientunitstayid,unitdischargestatus,er_los_hrs, sex, age, ethnicity, apachescoreIV, final_charlson_score, vasopressors, first_treatmentoffset) #hist_renal_bin, hist_diabetes_bin, hist_cancer_bin, hist_chf_bin, hist_copd_bin,hist_other_bin

final_df_er_sepsis <- distinct(final_df_er_sepsis)
```

```{r}
write.csv(final_df_er_sepsis, file = "C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Figures/selected_data.csv", row.names = FALSE)
```

# Flowchart

```{r}
titulo <- "Inclusion and Exclusion Criteria Flowchart"
# Crear el código DOT para el diagrama de flujo
flowchart_code <- sprintf("
digraph flowchart {
labelloc=\"t\";  // Colocar el título arriba del diagrama
  label=\"%s\";   //  Título del diagrama
  // Estilo del título
  graph [fontname = 'Arial', fontsize = 20, fontcolor = 'black', color = 'transparent'];
  node [shape = box, fontname = 'Arial', fontsize = 12, color = 'black'];
  a [label = 'Initial number of patients: %d'];
  b [label = 'Patients >=16 %d'];
  c [label = 'Patients in emergency department during 24 hrs: %d'];
  d [label = 'Patients with sex known: %d'];
  e [label = 'patient with emergency department: %d'];
  f [label = 'patient with unit stay type admit: %d'];
  g [label = 'Patients with apache admission dx known: %d'];
  h [label = 'Total Patients: %d']

  a -> b;
  b -> c;
  c -> d;
  d -> e;
  e -> f;
  f -> g;
  g -> h;
}
", titulo, a,b,c,d,e,f,g,h)

# Crear y visualizar el diagrama
grViz(flowchart_code)
```

# Subgroup analysis of septic shock (vasopressors)

```{r}
final_shock_séptico <- final_df_er_sepsis %>%
  select(hist_diabetes_bin, vasopressors) %>%
  filter(hist_diabetes_bin == 1)

final_df_er_sepsis <- distinct(final_shock_séptico)
```

# Boxplot time and diabetes

```{r}
ggplot(final_df_er_sepsis, aes(x = as.factor(hist_diabetes_bin), y = er_los_hrs)) +
  geom_boxplot() +
  labs(
    title = "Tiempo de Entrada en UCI según Diabetes",
    x = "Diabetes (0: No, 1: Sí)",
    y = "Tiempo en UCI (horas)"
  ) +
  theme_minimal()


final_df_er_sepsis %>% group_by(sex, hist_diabetes_bin) %>% summarise(n = n())
```



```{r}
## Correlation plot
# Get lower triangle of the correlation matrix
  get_lower_tri<-function(cormat){
    cormat[upper.tri(cormat)] <- NA
    return(cormat)
  }
  # Get upper triangle of the correlation matrix
  get_upper_tri <- function(cormat){
    cormat[lower.tri(cormat)]<- NA
    return(cormat)
  }
cormat <- round(cor(final_df_selected %>% select_if(is.numeric) %>%
                      cbind( final_df_selected %>% select_if(is.factor) %>% lapply(., as.numeric) %>% do.call(cbind, .)) %>% select(-icu_mortality_72hrs) %>% mutate(mortality = final_df_selected$icu_mortality_72hrs,
                                                           mortality = ifelse(mortality == "Alive", 0, 1))), 2)

final_df_selected %>% select_if(is.factor) %>% lapply(., as.numeric) %>% do.call(cbind, .) %>% nrow()
upper_tri <- get_upper_tri(cormat)
meted_cormat <- melt(upper_tri, na.rm = TRUE)



ggplot(data = meted_cormat, aes(Var2, Var1, fill = value))+
 geom_tile(color = "white")+
  geom_text(aes(label = value), col = "black",
            data = meted_cormat %>% filter(value > 0.09 | value < -0.09),
            size = 3) +
 scale_fill_gradient2(low = "blue", high = "red", mid = "white",
   midpoint = 0, limit = c(-1,1), space = "Lab",
   name="Pearson\nCorrelation") +
  theme_minimal()+
 theme(axis.text.x = element_text(angle = 45, vjust = 1,
    size = 12, hjust = 1))+
 coord_fixed()
```
