==> Analysis/03.tables.Rmd <==
---
title: "03_tables"
output: html_document
---

# Environment
```{r}
library(tableone)
library(kableExtra)
library(knitr)
library(janitor)
library(Hmisc)
library(htmltools)

```

#Table one 1º

```{r}
final_df_er_sepsis <- read.csv("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Figures/selected_data.csv")
message("Datos cargados correctamente")

#List of variables included in table 1

vars_in_table1 <- names(final_df_er_sepsis %>% select(unitdischargestatus,er_los_hrs, sex, age, ethnicity, apachescoreIV, final_charlson_score, vasopressors,first_treatmentoffset))

```

#Create a dataset with only selected variables using our original dataset and generate an empty list to store categorical variables

```{r}
table1_dataset <- final_df_er_sepsis[,vars_in_table1]
cat_variables <- rep(NA, length(vars_in_table1))
stratifyby <- "unitdischargestatus"

```

#Detects whether a variable is categorical or not

```{r}
cont <- 1
for (i in 1:length(vars_in_table1)) {
  if ( n_distinct(table1_dataset[vars_in_table1[i]]) <= 10) {
    print(i)
    print(vars_in_table1[i])
    print(names(table1_dataset[vars_in_table1[i]]))
    cat_variables[cont] <- names(table1_dataset[vars_in_table1[i]])
    cont <- cont+1
  }
} 
cat_variables <- cat_variables[!is.na(cat_variables)]
table1_base <- print(CreateTableOne(vars = setdiff(vars_in_table1, stratifyby)
                                    , strata = stratifyby
                                    , factorVars = cat_variables
      ,data = table1_dataset, addOverall = T), varLabels = T)

stratification_cats <- n_distinct(table1_dataset[,stratifyby])

table1_base %>%
  kbl(caption = "Table 1. Emergency and sepsis. Stratified by ICU mortality", align = "c") %>%
  kable_classic_2(full_width = F, html_font = "Cambria") %>%
  add_header_above(c("","", "Alived", "Expired","",""))

# Export table
table1_base %>%
  kbl(caption = "Table 1. Emergency and sepsis. Stratified by ICU mortality", align = "c") %>%
  kable_classic_2(full_width = F, html_font = "Cambria") %>%
  add_header_above(c("","","Alived","Expired","","")) %>%
  save_kable("C:/Users/cjimenez10J/Documents/projects/r_urgencias_sepsis/r_urgencias_sepsis/Figures/table1_base.html", zoom = 10)


```
